Move TriggeredServerCodeResultTest.
